requirements/base.txt
Django==1.7.5
Pillow==2.8.1
django-absolute==0.3
django-braces==1.8.0
django-haystack==2.3.1
django-model-utils==2.2
django-recaptcha==1.0.2
django-uuidfield==0.5.0
djangorestframework==3.0.2
djrill==1.3.0
easy-thumbnails==2.2
elasticsearch==1.4.0
mock==1.0.1
model-mommy==1.2.2
ply==3.4
psycopg2==2.6
pyelasticsearch==0.7.1
python-memcached==1.53
requests==2.5.1
simplejson==3.6.5
six==1.7.3
slimit==0.8.1
sqlparse==0.1.14
urllib3==1.10.2
wsgiref==0.1.2